smd: fix sign issue in uint64_t >= int64_t when latter is negative

+++ module/spl/spl-kmem.c
@@ -4186,7 +4186,7 @@ spl_free_thread()
 			spl_free -= (pct_used - 97) * (int64_t)(tunable_osif_memory_cap / 1000ULL);
 
 			lowmem = true;
-		} else if (segkmem_total_mem_allocated >= tunable_osif_memory_cap) {
+		} else if (segkmem_total_mem_allocated >= (int64_t)tunable_osif_memory_cap) {
 			if (spl_free > (16LL * 1024LL * 1024LL))    // SPA_MAXBLOCKSIZE
 				spl_free = 16LL * 1024LL * 1024LL;
 		}
